feat: separate `wrapMethod` via `wrap-transaction`
- to test transaction, it has to separate from explorer.
(because it has tight coupling with nest-js)
